+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Avoid invariant failure when updating the status of a `AreaOfInterest` for a subscription that
   is implicitly subscribed and is in the process of being explicitly subscribed but is currently
   in `NOT_ASKED` or `LOADING` state.
+* Avoid dropping subscription requests in the `Connector` when the request is upgrading an
+  implicit subscription to an explicit subscription.
 
 ### [v6.27](https://github.com/realityforge/replicant/tree/v6.27) (2019-02-04)
 [Full Changelog](https://github.com/realityforge/replicant/compare/v6.26...v6.27)